+++ src/synchronous.rs
@@ -1,0 +1,310 @@
+use embedded_hal as hal;
+use hal::delay::DelayNs;
+use hal::i2c::I2c;
+
+use sensirion_i2c::{crc8, i2c};
+
+use crate::{Command, Error};
+#[cfg(feature = "voc_index")]
+use crate::vocalg::VocAlgorithm;
+
+
+/// Sgp40 driver instance
+///
+/// Create the driver instance with valid I²C address (0x59) and then it is just
+/// rock'n'roll. This driver doesn't require special starting but once can start to
+/// make measurements right away. However, the initial values after start-up will
+/// unstable so you will want to throw away some of them.
+pub struct Sgp40<I2C, D> {
+    i2c: I2C,
+    address: u8,
+    delay: D,
+    temperature_offset: i16,
+    #[cfg(feature = "voc_index")]
+    voc: VocAlgorithm,
+}
+
+impl<I2C, D, E> Sgp40<I2C, D>
+where
+    I2C: hal::i2c::I2c<Error = E>,
+    D: DelayNs,
+{
+    /// Creates Sgp40 driver
+    pub fn new(i2c: I2C, address: u8, delay: D) -> Self {
+        Sgp40 {
+            i2c,
+            address,
+            delay,
+            temperature_offset: 0,
+            #[cfg(feature = "voc_index")]
+            voc: VocAlgorithm::new(),
+        }
+    }
+
+    /// Command for reading values from the sensor
+    fn delayed_read_cmd(&mut self, cmd: Command, data: &mut [u8]) -> Result<(), Error<E>> {
+        self.write_command(cmd)?;
+        i2c::read_words_with_crc(&mut self.i2c, self.address, data)?;
+        Ok(())
+    }
+
+    /// Writes commands with arguments
+    fn write_command_with_args(&mut self, cmd: Command, data: &[u8]) -> Result<(), Error<E>> {
+        const MAX_TX_BUFFER: usize = 14; //cmd (2 bytes) + max args (12 bytes)
+
+        let mut transfer_buffer = [0; MAX_TX_BUFFER];
+
+        let size = data.len();
+
+        // 2 for command, size of transferred bytes and CRC per each two bytes.
+        assert!(size < 2 + size + size / 2);
+        let (command, delay) = cmd.as_tuple();
+
+        transfer_buffer[0..2].copy_from_slice(&command.to_be_bytes());
+
+        let mut i = 2;
+        for chunk in data.chunks(2) {
+            let end = i + 2;
+            transfer_buffer[i..end].copy_from_slice(chunk);
+            transfer_buffer[end] = crc8::calculate(chunk);
+            i += 3;
+        }
+
+        self.i2c
+            .write(self.address, &transfer_buffer[0..i])
+            .map_err(Error::I2c)?;
+        self.delay.delay_ms(delay);
+
+        Ok(())
+    }
+
+    /// Writes commands without additional arguments.
+    fn write_command(&mut self, cmd: Command) -> Result<(), Error<E>> {
+        let (command, delay) = cmd.as_tuple();
+        i2c::write_command_u16(&mut self.i2c, self.address, command).map_err(Error::I2c)?;
+        self.delay.delay_ms(delay);
+        Ok(())
+    }
+
+    /// Sensor self-test.
+    ///
+    /// Performs sensor self-test. This is intended for production line and testing and verification only and
+    /// shouldn't be needed for normal use.
+    pub fn self_test(&mut self) -> Result<&mut Self, Error<E>> {
+        const MEASURE_TEST_OK: u16 = 0xd400;
+        let mut data = [0; 3];
+
+        self.delayed_read_cmd(Command::MeasureTest, &mut data)?;
+
+        let result = u16::from_be_bytes([data[0], data[1]]);
+
+        if result != MEASURE_TEST_OK {
+            Err(Error::SelfTest)
+        } else {
+            Ok(self)
+        }
+    }
+
+    /// Turn sensor heater off and places it in idle-mode.
+    ///
+    /// Stops running the measurements, places heater into idle by turning the heaters off.
+    #[inline]
+    pub fn turn_heater_off(&mut self) -> Result<&Self, Error<E>> {
+        self.write_command(Command::HeaterOff)?;
+        Ok(self)
+    }
+
+    /// Resets the sensor.
+    ///
+    /// Executes a reset on the device. The caller must wait 100ms before starting to use the device again.
+    #[inline]
+    pub fn reset(&mut self) -> Result<&Self, Error<E>> {
+        self.write_command(Command::SoftReset)?;
+        Ok(self)
+    }
+
+    /// Reads the voc index from the sensor.
+    ///
+    /// Reads VOC index. Driver is using Sensirion proprietary algortihm and it takes minimum
+    /// 45 reads to start working. These reads should be made with 1Hz interval to keep the
+    /// algoritm working.
+    #[cfg(feature = "voc_index")]
+    #[inline]
+    pub fn measure_voc_index(&mut self) -> Result<u16, Error<E>> {
+        let raw = self.measure_raw_with_rht(50000, 25000)?;
+
+        Ok(self.voc.process(raw as i32) as u16)
+    }
+
+    /// Reads the voc index from the sensor with humidity and temperature compensation.
+    ///
+    /// Reads VOC index with humidity and temperature compensation. Both values us milli-notation where
+    /// 25°C is equivalent of 25000 and 50% humidity equals 50000.
+    ///
+    /// Driver is using Sensirion proprietary algortihm and it takes minimum
+    /// 45 reads to start working. These reads should be made with 1Hz interval to keep the
+    /// algoritm working.
+    #[cfg(feature = "voc_index")]
+    #[inline]
+    pub fn measure_voc_index_with_rht(&mut self, humidity: u16, temperature: i16) -> Result<u16, Error<E>> {
+        let raw = self.measure_raw_with_rht(humidity, temperature)?;
+
+        Ok(self.voc.process(raw as i32) as u16)
+    }
+
+    /// Reads the raw signal from the sensor.
+    ///
+    /// Raw signal without temperature and humidity compensation. This is not
+    /// VOC index but needs to be processed through different algorithm for that.
+    #[inline]
+    pub fn measure_raw(&mut self) -> Result<u16, Error<E>> {
+        self.measure_raw_with_rht(50000, 25000)
+    }
+
+    /// Reads the raw signal from the sensor.
+    ///
+    /// Raw signal with temperature and humidity compensation. This is not
+    /// VOC index but needs to be processed through different algorithm for that.
+    pub fn measure_raw_with_rht(&mut self, humidity: u16, temperature: i16) -> Result<u16, Error<E>> {
+        let mut data = [0; 3];
+
+        let (hum_ticks, temp_ticks) = self.convert_rht(humidity as u32, temperature as i32);
+
+        let mut params = [0u8; 4];
+        params[0..2].copy_from_slice(&hum_ticks.to_be_bytes());
+        params[2..4].copy_from_slice(&temp_ticks.to_be_bytes());
+
+        self.write_command_with_args(Command::MeasurementRaw, &params)?;
+        i2c::read_words_with_crc(&mut self.i2c, self.address, &mut data)?;
+
+        Ok(u16::from_be_bytes([data[0], data[1]]))
+    }
+
+    // Returns tick converted values
+    fn convert_rht(&self, humidity: u32, temperature: i32) -> (u16, u16) {
+        let mut temperature = temperature;
+        let mut humidity = humidity;
+        if humidity > 100000 {
+            humidity = 100000;
+        }
+
+        temperature += self.temperature_offset as i32;
+
+        temperature = temperature.clamp(-45000, 129760);
+
+        /* humidity_sensor_format = humidity / 100000 * 65535;
+         * 65535 / 100000 = 0.65535 -> 0.65535 * 2^5 = 20.9712 / 2^10 ~= 671
+         */
+        let humidity_sensor_format = ((humidity * 671) >> 10) as u16;
+
+        /* temperature_sensor_format[1] = (temperature + 45000) / 175000 * 65535;
+         * 65535 / 175000 ~= 0.375 -> 0.375 * 2^3 = 2.996 ~= 3
+         */
+        let temperature_sensor_format = (((temperature + 45000) * 3) >> 3) as u16;
+
+        (humidity_sensor_format, temperature_sensor_format)
+    }
+
+    /// Sets the temperature offset.
+    ///
+    /// This command sets the temperature offset used for the compensation of subsequent RHT measurements.
+    /// The parameter provides the temperature offset (in °C) with a scaling factor of 200, e.g., an output of +400 corresponds to +2.00 °C.
+    #[inline]
+    pub fn set_temperature_offset(&mut self, offset: i16) -> Result<&mut Self, Error<E>> {
+        self.temperature_offset += offset;
+        Ok(self)
+    }
+
+    /// Gets the temperature offset
+    ///
+    /// Gets the temperature compensation offset issues to the device.
+    pub fn get_temperature_offset(&mut self) -> Result<i16, Error<E>> {
+        Ok(self.temperature_offset)
+    }
+
+    /// Acquires the sensor serial number.
+    ///
+    /// Sensor serial number is only 48-bits long so the remaining 16-bits are zeros.
+    pub fn serial(&mut self) -> Result<u64, Error<E>> {
+        let mut serial = [0; 9];
+
+        self.delayed_read_cmd(Command::Serial, &mut serial)?;
+
+        let serial = (u64::from(serial[0]) << 40)
+            | (u64::from(serial[1]) << 32)
+            | (u64::from(serial[3]) << 24)
+            | (u64::from(serial[4]) << 16)
+            | (u64::from(serial[6]) << 8)
+            | u64::from(serial[7]);
+        Ok(serial)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use embedded_hal_mock as mock_hal;
+
+    use super::*;
+
+    use mock_hal::eh1::{
+        delay::NoopDelay,
+        i2c::{Mock as I2cMock, Transaction},
+    };
+
+    const SGP40_ADDR: u8 = 0x59;
+
+    /// Tests that the commands without parameters work
+    #[test]
+    fn test_basic_command() {
+        let (cmd, _) = Command::MeasurementRaw.as_tuple();
+        let expectations = [
+            Transaction::write(
+                SGP40_ADDR,
+                [
+                    cmd.to_be_bytes().to_vec(),
+                    [0x7f, 0xfb, 0x4b, 0x66, 0x8a, 0x2f].to_vec(),
+                ]
+                .concat(),
+            ),
+            Transaction::read(SGP40_ADDR, vec![0x12, 0x34, 0x37]),
+        ];
+        let mut mock = I2cMock::new(&expectations);
+        let mut sensor = Sgp40::new(mock.clone(), SGP40_ADDR, NoopDelay);
+        let result = sensor.measure_raw().unwrap();
+        assert_eq!(result, 0x1234);
+        mock.done();
+    }
+
+    /// Test the `serial` function
+    #[test]
+    fn serial() {
+        let (cmd, _) = Command::Serial.as_tuple();
+        let expectations = [
+            Transaction::write(0x58, cmd.to_be_bytes().to_vec()),
+            Transaction::read(0x58, vec![0xde, 0xad, 0x98, 0xbe, 0xef, 0x92, 0xde, 0xad, 0x98]),
+        ];
+        let mut mock = I2cMock::new(&expectations);
+        let mut sensor = Sgp40::new(mock.clone(), 0x58, NoopDelay);
+        let serial = sensor.serial().unwrap();
+        assert_eq!(serial, 0x00deadbeefdead);
+        mock.done();
+    }
+
+    #[test]
+    fn test_crc_error() {
+        let (cmd, _) = Command::MeasureTest.as_tuple();
+        let expectations = [
+            Transaction::write(SGP40_ADDR, cmd.to_be_bytes().to_vec()),
+            Transaction::read(SGP40_ADDR, vec![0xD4, 0x00, 0x00]),
+        ];
+        let mut mock = I2cMock::new(&expectations);
+        let mut sensor = Sgp40::new(mock.clone(), SGP40_ADDR, NoopDelay);
+
+        match sensor.self_test() {
+            Err(Error::Crc) => {}
+            Err(_) => panic!("Unexpected error in CRC test"),
+            Ok(_) => panic!("Unexpected success in CRC test"),
+        }
+        mock.done();
+    }
+}
